requirements.txt
pytest>=7.0
pytest-cov>=4.0
requests>=2.28
feedparser>=6.0
openai>=1.0
ruff>=0.1
python-dotenv>=1.1
